~ Initial layered fog added;

+++ glsl/vs-depthmap.glsl
@@ -1,4 +1,5 @@
-precision mediump float;
+precision highp float;
+
 attribute vec3 aVertexPosition;
 
 uniform mat4 uMMatrix;
